qr_generator/requirements.txt
qrcode[pil]==7.4.2
pillow==10.3.0